pyproject.toml
[project]
name = "dcrsystem-pc-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hidapi>=0.14.0.post4",
    "opencv-python>=4.12.0.88",
    "pypylon>=4.2.0",
    "pyside6[tools]>=6.10.1",
    "requests>=2.32.5",
]
